Removed Entry from code and handled a few exceptions
