+++ app/src/main/java/com/cmput301f18t20/medicalphotorecord/Problem.java
@@ -5,13 +5,16 @@ import android.location.Location;
 import java.util.ArrayList;
 import java.util.Date;
 
-public class Problem extends Entry implements Refreshable {
+public class Problem implements Refreshable {
     protected String description;
     protected ArrayList<Record> records = new ArrayList<>();
+    protected String title, createdByUserID;
+    protected Date date = new Date(System.currentTimeMillis());
 
     public Problem(String createdByUserID, String title)
             throws UserIDMustBeAtLeastEightCharactersException, TitleTooLongException {
-        super(createdByUserID, title);
+        this.setCreatedByUserID(createdByUserID);
+        this.setTitle(title);
     }
 
     public Record getRecord(int recordIndex) {
@@ -109,5 +112,37 @@ public class Problem extends Entry implements Refreshable {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    public String getCreatedByUserID() {
+        return createdByUserID;
+    }
+
+    public void setCreatedByUserID(String createdByUserID)
+            throws UserIDMustBeAtLeastEightCharactersException {
+        if (createdByUserID.length() >= 8) {
+            this.createdByUserID = createdByUserID;
+        } else {
+            throw new UserIDMustBeAtLeastEightCharactersException();
+        }
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) throws TitleTooLongException {
+        if (title.length() > 30) {
+            throw new TitleTooLongException();
+        }
+        this.title = title;
+    }
+
+    public Date getDate() {
+        return date;
+    }
+
+    public void setDate(Date date) {
+        this.date = date;
     }
 }
