+++ app/src/main/java/com/cmput301f18t20/medicalphotorecord/Record.java
@@ -4,12 +4,14 @@ import android.location.Location;
 
 import java.util.Date;
 
-public class Record extends Entry {
-    String comment;
+public class Record {
+    protected String comment, title, createdByUserID;
+    protected Date date = new Date(System.currentTimeMillis());
 
     Record(String creatorUserID, String title)
             throws UserIDMustBeAtLeastEightCharactersException, TitleTooLongException {
-        super(creatorUserID, title);
+        this.setCreatedByUserID(creatorUserID);
+        this.setTitle(title);
     }
 
     public String getComment() {
@@ -21,5 +23,37 @@ public class Record extends Entry {
             throw new CommentTooLongException();
         }
         this.comment = comment;
+    }
+
+    public String getCreatedByUserID() {
+        return createdByUserID;
+    }
+
+    public void setCreatedByUserID(String createdByUserID)
+            throws UserIDMustBeAtLeastEightCharactersException {
+        if (createdByUserID.length() >= 8) {
+            this.createdByUserID = createdByUserID;
+        } else {
+            throw new UserIDMustBeAtLeastEightCharactersException();
+        }
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) throws TitleTooLongException {
+        if (title.length() > 30) {
+            throw new TitleTooLongException();
+        }
+        this.title = title;
+    }
+
+    public Date getDate() {
+        return date;
+    }
+
+    public void setDate(Date date) {
+        this.date = date;
     }
 }
