+++ app/src/main/java/com/cmput301f18t20/medicalphotorecord/UserIDMustBeAtLeastEightCharactersException.java
@@ -1,4 +1,7 @@
 package com.cmput301f18t20.medicalphotorecord;
 
 public class UserIDMustBeAtLeastEightCharactersException extends Exception {
+    public UserIDMustBeAtLeastEightCharactersException() {
+        super("User id needs to be at least 8 characters");
+    }
 }
